Summary:
Merge pg15 commit 4a5799c93dc3b7aa7221fd755567b64fc42c8681 titled

    [pg15] test: get TestPgRegressTypesGeo to pass

 - test_lint.sh: Merge change in list of exceptions to lint rules
 - pg_yb_common.{c,h}: Lines added in the same place. Take both changes.
 - pg_session.cc: Initializers were reordered because fields were reordered in pg_session.h, and the initializer for buffer_ was rewritten. Simple resolution.
 - pg_wrapper.{cc,h}: pg15-upgrade branch commit d25d78c3026c2c0481038e8065f0d2329e1ab296 makes further modifications for database OID parity, on top of versioned_data_dir changes for local initdb in master commit 2b7b8de3dd0bbe925478b5872fbc19c3e331d6f2. Take the more recent changes from the pg15-upgrade branch.

Test Plan:
Jenkins: rebase: pg15-upgrade

On MacOS 13 arm64:
  ./yb_build.sh release --sj
  pg15_tests/test_upgrade.sh
  pg15_tests/test_upgrade_dropped_col.sh

On AlmaLinux 8:
  ./yb_build.sh fastdebug --gcc11
  pg15_tests/run_tests.sh
test_upgrade.sh and test_upgrade_dropped_col.sh will fail because release and debug builds cannot interoperate.
Also, there's a flaky test fixed in in 528164d778.

Reviewers: jason

Reviewed By: jason

Subscribers: esheng, bogdan, yql, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D33406

+++ managed/src/main/java/com/yugabyte/yw/forms/DrConfigSafetimeResp.java
@@ -21,9 +21,9 @@ public class DrConfigSafetimeResp {
     private final long safetimeEpochUs;
     private final long safetimeLagUs;
     private final long safetimeSkewUs;
-    private final long estimatedDataLossMs;
+    private final double estimatedDataLossMs;
 
-    public NamespaceSafetime(NamespaceSafeTimePB namespaceSafeTimePB, long estimatedDataLossMs) {
+    public NamespaceSafetime(NamespaceSafeTimePB namespaceSafeTimePB, double estimatedDataLossMs) {
       this.namespaceId = namespaceSafeTimePB.getNamespaceId();
       this.namespaceName = namespaceSafeTimePB.getNamespaceName();
       this.safetimeEpochUs =
